Provide option to redirect program output to target
Currently the program redirected output for a remote deploy is
written to the host machine. This has the benefit of getting the
log and errors back before a target may crash. The downside
is that it can send a lot of data across the remote connection
(adb). We are observing the adb process crashing on the device
when large amounts of data is redirected back to the host. The
solution for us is to write the redirected output to the
device and have it treated like a normal run output file.